requirements.txt
alembic==0.9.7
backports.functools-lru-cache==1.6.1
bcolz==1.2.1
boto3==1.5.27
botocore==1.8.50
Bottleneck==1.2.1
ccxt==1.17.94
certifi==2018.1.18
chardet==3.0.4
click==6.7
contextlib2==0.5.5
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.27.3
decorator==4.3.0
docutils==0.16
empyrical==0.2.1
enigma-catalyst==0.5.21
futures==3.3.0
idna==2.7
intervaltree==2.1.0
jmespath==0.9.5
kiwisolver==1.1.0
Logbook==0.12.5
lru-dict==1.1.6
lxml==4.5.0
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.5
multipledispatch==0.4.9
networkx==2.1
numexpr==2.6.4
numpy==1.14.0
pandas==0.19.2
pandas-datareader==0.6.0
patsy==0.5.0
pyparsing==2.4.6
python-dateutil==2.7.3
python-editor==1.0.4
pytz==2016.4
redo==2.0.1
requests==2.20.1
requests-file==1.4.3
requests-ftp==0.3.1
requests-toolbelt==0.8.0
s3transfer==0.1.13
scipy==1.0.0
six==1.11.0
sortedcontainers==1.5.9
SQLAlchemy==1.2.2
statsmodels==0.8.0
tables==3.4.2
toolz==0.9.0
urllib3==1.24.3
wincertstore==0.2
wrapt==1.12.1
